fix(simulation_objects.h): change calories type from int to float to match food type

+++ src/simulation/utilities/GeneralFunctions.h
@@ -12,7 +12,7 @@ constexpr float caloriesToSizeKoef = 1.0f;
 /// @param calories Amount of calories
 /// @param koef Koef to multiply radius by
 /// @return Radius of circle containing given amount of calories
-inline int convertCaloriesToRadius(int calories, float koef = caloriesToSizeKoef)
+inline int convertCaloriesToRadius(float calories, float koef = caloriesToSizeKoef)
 {
     return int(ceil(sqrt(calories / M_PI) * koef));
 }
